success in separating stylesheet and applying to pages. ran the DRYer
