+++ .expo/packager-info.json
@@ -2,8 +2,8 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 2399,
-  "expoServerNgrokUrl": "https://pi-3st.anonymous.neighbor-corps-react-native.exp.direct",
-  "packagerNgrokUrl": "https://packager.pi-3st.anonymous.neighbor-corps-react-native.exp.direct",
-  "ngrokPid": 2414
+  "packagerPid": 3597,
+  "expoServerNgrokUrl": null,
+  "packagerNgrokUrl": null,
+  "ngrokPid": null
 }
