+++ screens/LoginScreen.js
@@ -1,24 +1,37 @@
 import React, {Component} from 'react';
 import { ExpoConfigView } from '@expo/samples';
-import { Container, Header, Content, Form, Item, Input, Label, Text } from 'native-base';
+import { Container, Content, Item, Input, Label, Text } from 'native-base';
+import { TouchableHighlight } from 'react-native';
+import Header from '../components/Header';
+
+const reactStyles = require('../react_native_styles/styles');
+const styles = reactStyles.default;
+
+import t from 'tcomb-form-native';
+
+const Form = t.form.Form;
+
+const Login = t.struct({
+  username: t.String,
+  password: t.String,
+});
 
 export default class LoginScreen extends Component {
+  state = {
+    page: "Log In"
+  }
   render() {
     return (
       <Container>
-        <Header />
+        <Header page={this.state.page}/>
         <Content>
-        <Text>SIGN UP SCREEN</Text>
-          <Form>
-            <Item floatingLabel>
-              <Label>Username</Label>
-              <Input />
-            </Item>
-            <Item floatingLabel last>
-              <Label>Password</Label>
-              <Input />
-            </Item>
-          </Form>
+        <Form
+            ref="form"
+            type={Login}
+          />
+          <TouchableHighlight style={styles.button} onPress={this.onPress} underlayColor='#99d9f4'>
+            <Text style={styles.buttonText}>Log In</Text>
+          </TouchableHighlight>
         </Content>
       </Container>
     );
